add PropTypes for footer
